updated the PyPoll main.py, PyPoll.txt and screenshot of results
An interesting result. The python program can access all the data from the spreadsheets if it is not opened first. I became aware of the great amount of data when I read the tips of the project. I noticed the election_data_2 was actually 4 times the amount of megabytes as the election_data_1. When I opened it, it would break and go to the max records on the sheet. When I exported to notepad and saw how many were actually there, I thought to try to copy and paste. I ran my program on the closed files and the numbers came out right. Very Interesting.