Best attempt to do more
The two challenges I completed were PyBank and PyBoss. The files submitted here under PyBoss show my attempt to go beyond the required and give a view of where I need to learn more. I included a number of files to give a full picture of my attempts. My challenges were mainly in the use of strptime and date conversion to the DD-MM-YYYY and the SSN masking. Comments to help me learn are greatly appreciated.